# internal/db/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Installation struct {
	ID             pgtype.UUID        `json:"id"`
	Name           string             `json:"name"`
	ClientMode     string             `json:"client_mode"`
	Enabled        bool               `json:"enabled"`
	Policy         []byte             `json:"policy"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
	ResolutionMode string             `json:"resolution_mode"`
	ResolverID     pgtype.UUID        `json:"resolver_id"`
	PublicToken    string             `json:"public_token"`
}

type Provider struct {
	ID               pgtype.UUID        `json:"id"`
	Name             string             `json:"name"`
	Kind             string             `json:"kind"`
	Endpoint         string             `json:"endpoint"`
	Enabled          bool               `json:"enabled"`
	Config           []byte             `json:"config"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
	SecretCiphertext []byte             `json:"secret_ciphertext"`
	SecretNonce      []byte             `json:"secret_nonce"`
}

type ResolverAccount struct {
	ID               pgtype.UUID        `json:"id"`
	Name             string             `json:"name"`
	Kind             string             `json:"kind"`
	Enabled          bool               `json:"enabled"`
	SecretCiphertext []byte             `json:"secret_ciphertext"`
	SecretNonce      []byte             `json:"secret_nonce"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
}